Renames asm packages to match folder name
