+++ internal/asm/arm64/impl.go
@@ -1,4 +1,4 @@
-package asm_arm64
+package arm64
 
 import (
 	"bytes"
